packages/sweetpotato/sweetpotato-0.7.0a0.tar.gz/sweetpotato-0.7.0a0/src/tests/test_backend/test_app.py
"""Unittests for App class."""
import unittest


class TestApp(unittest.TestCase):

    def setUp(self) -> None:
        """Set up app."""
        # component = View(children=[Text(text="Hello, World")])
        # self.app_repr = '<SafeAreaProvider ><View ><Text >Hello, World</Text></View></SafeAreaProvider>'
        # self.app = App(component=component)

    def test_props(self):
        """
        Test that App class props are equal to components_props.py variable.
        """
        ...

    def test_has_attributes(self):
        ...

    def test_no_component(self):
        ...

    def test_run(self):
        ...
        # self.assertTrue(self.app.run)

    def test_publish(self):
        # self.assertTrue(self.app.publish)
        ...

    def test_show(self):
        ...
        # self.assertIsInstance(str(self.app.show()), str)
        # self.assertEqual(str(self.app.show()), self.app_repr)


if __name__ == "__main__":
    unittest.main()
